main-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
google-genai==0.3.0
pydantic==2.6.3
pydantic-settings==2.2.1
requests==2.31.0
httpx==0.27.0
